[refact] 좌석 배치 영역 컴포넌트 분리
